+++ vortex-cuda/src/lib.rs
@@ -9,8 +9,7 @@ mod canonical;
 mod device_buffer;
 pub mod executor;
 mod kernel;
-mod pinned;
-mod pinned_allocator;
+mod host_to_device_allocator;
 mod session;
 mod stream;
 
@@ -19,11 +18,7 @@ pub use device_buffer::CudaBufferExt;
 pub use device_buffer::CudaDeviceBuffer;
 pub use executor::CudaExecutionCtx;
 pub use executor::CudaKernelEvents;
-pub use pinned::PinnedByteBuffer;
-pub use pinned::PinnedByteBufferPool;
-pub use pinned::PooledPinnedBuffer;
-pub use pinned_allocator::PinnedBufferAllocator;
-pub use pinned_allocator::PinnedDeviceAllocator;
+pub use host_to_device_allocator::HostToDeviceAllocator;
 use kernel::ALPExecutor;
 use kernel::DecimalBytePartsExecutor;
 use kernel::DictExecutor;
